% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/character-quotes.R
\name{character_quotes}
\alias{character_quotes}
\alias{michael}
\alias{jim}
\alias{dwight}
\alias{pam}
\alias{andy}
\alias{kevin}
\alias{angela}
\alias{erin}
\alias{oscar}
\alias{ryan}
\alias{darryl}
\alias{phyllis}
\alias{toby}
\alias{kelly}
\alias{stanley}
\alias{meredith}
\alias{creed}
\title{Get quotes from specific characters}
\usage{
michael(...)

jim(...)

dwight(...)

pam(...)

andy(...)

kevin(...)

angela(...)

erin(...)

oscar(...)

ryan(...)

darryl(...)

phyllis(...)

toby(...)

kelly(...)

stanley(...)

meredith(...)

creed(...)
}
\arguments{
\item{...}{arguments to be passed to \link[dundermifflin]{get_quote}}
}
\description{
Get quotes from specific characters
}
\examples{
\dontrun{
  michael()
  jim()
  kelly()
  dwight()
}


}
