% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/departments.R
\name{department_quotes}
\alias{department_quotes}
\alias{sales}
\alias{accounting}
\alias{customer_service}
\alias{hr}
\alias{reception}
\alias{corporate}
\title{Get quotes from different departments of the office}
\usage{
sales(...)

accounting(...)

customer_service(...)

hr(...)

reception(...)

corporate(...)
}
\arguments{
\item{...}{arguments to be passed to \link[dundermifflin]{get_quote}}
}
\description{
Get quotes from different departments of the office
}
\examples{
\dontrun{
  sales()
  accounting()
}


}
