% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-quotes.R
\name{filter_quotes}
\alias{filter_quotes}
\title{Filter office_quotes}
\usage{
filter_quotes(season = NULL, episode = NULL, name = NULL,
  scene = NULL, character = "main", min_nword = 10,
  max_nword = 100, include_actions = FALSE)
}
\arguments{
\item{season}{an integer specifyng the season for which to get quotes from
Currently, all seasons are included except season 5}

\item{episode}{an integer specifying the episode number to get quote from}

\item{name}{a character string with the episode name to get quote from}

\item{scene}{an integer specifying the scene number}

\item{character}{a character specifying the character or department to get quote from.
Defaults to 'main', see details}

\item{min_nword}{the minimum number of words a quote must have to be
included, defaults to 4}

\item{max_nword}{the maximum number of words a quote can have to be included,
defaults to 20}

\item{include_actions}{should non spoken actions be included in the output?
Defaults to FALSE (i.e. actions will not be included)}
}
\description{
Filter office_quotes
}
\details{
Main Character List - Michael, Dwight, Jim, Pam, Andy, Kevin, Angela,
Erin, Oscar, Ryan, Darryl, Phyllis, Toby, Kelly, Stanley, Meredith,
and Creed

Other Departments - Sales, Accounting, HR, Customer Service, Reception,
and Corporate
}
