% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{office_quotes}
\alias{office_quotes}
\alias{ep_n_scene}
\title{Dataset of all quotes from The Office}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 50221 rows and 7 columns.}
\usage{
office_quotes

ep_n_scene
}
\description{
Dataset of all quotes from The Office

Season, episode and scene combinations
}
\keyword{datasets}
